+++ lib/printing_ffi_bindings_generated.dart
@@ -105,6 +105,7 @@ class PrintingFfiBindings {
     ffi.Pointer<ffi.Char> printer_name,
     ffi.Pointer<ffi.Char> pdf_file_path,
     ffi.Pointer<ffi.Char> doc_name,
+    int scaling_mode,
     int num_options,
     ffi.Pointer<ffi.Pointer<ffi.Char>> option_keys,
     ffi.Pointer<ffi.Pointer<ffi.Char>> option_values,
@@ -113,14 +114,15 @@ class PrintingFfiBindings {
       printer_name,
       pdf_file_path,
       doc_name,
+      scaling_mode,
       num_options,
       option_keys,
       option_values,
     );
   }
 
-  late final _print_pdfPtr = _lookup<ffi.NativeFunction<ffi.Bool Function(ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, ffi.Int, ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Pointer<ffi.Char>>)>>('print_pdf');
-  late final _print_pdf = _print_pdfPtr.asFunction<bool Function(ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, int, ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Pointer<ffi.Char>>)>();
+  late final _print_pdfPtr = _lookup<ffi.NativeFunction<ffi.Bool Function(ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, ffi.Int, ffi.Int, ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Pointer<ffi.Char>>)>>('print_pdf');
+  late final _print_pdf = _print_pdfPtr.asFunction<bool Function(ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, ffi.Pointer<ffi.Char>, int, int, ffi.Pointer<ffi.Pointer<ffi.Char>>, ffi.Pointer<ffi.Pointer<ffi.Char>>)>();
 
   ffi.Pointer<JobList> get_print_jobs(
     ffi.Pointer<ffi.Char> printer_name,